retry deployments in all apps
